menambahkan situasi pada PUT, POST, DELETE

+++ src/main/java/pws/tutorialspoint/ProductServiceController.java
@@ -43,23 +43,45 @@ public class ProductServiceController {
     }
     @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Object> delete(@PathVariable("id") String id) {
-        productRepo.remove(id);
+        if(!productRepo.containsKey(id))
+        {
+            return new ResponseEntity<>("Product missing", HttpStatus.NOT_FOUND);
+        }
+        else
+        {
+            productRepo.remove(id);
         return new ResponseEntity<> ("Product is delete success", HttpStatus.OK);
+        }
     }
         @RequestMapping (value = "/products/{id}", method = RequestMethod.PUT)
         public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product){
+            if (!productRepo.containsKey(id))
+            {
+             return new ResponseEntity<>("Product missing", HttpStatus.NOT_FOUND);
+            }
+            else
+            {
             productRepo.remove(id);
             product.setId(id);
             productRepo.put(id, product);
             return new ResponseEntity<>("Product update success", HttpStatus.OK);
-            
+            } 
             
         }
         @RequestMapping(value = "/products", method = RequestMethod.POST)
         public ResponseEntity<Object> createProduct(@RequestBody Product product){
+            if(productRepo.containsKey(product.getId()))
+            {
+                return new ResponseEntity<>("Product missing", HttpStatus.NOT_FOUND);
+                
+            }
+            else
+            {
+            
+            
             productRepo.put(product.getId(), product);
             return new ResponseEntity<>("Product created succes", HttpStatus.OK);
-            
+            }
         
         }
         @RequestMapping(value = "/products")
